Extract shared crate with UserData and message enums
Move UserData, ClientRequestMessage, and ClientResponseMessage into
the shared crate. Add actix Message derives behind an optional feature
flag. Remove data.rs re-export files from server and frontend.
